TwitterSentiment/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asn1crypto=0.24.0=py36_1003
blinker=1.4=py_1
ca-certificates=2019.1.23=0
certifi=2019.3.9=py36_0
cffi=1.12.2=py36h2d6ddff_1
chardet=3.0.4=py36_1003
cryptography=2.6.1=py36h212c5bf_0
dnspython=1.16.0=py36_0
idna=2.8=py36_1000
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
ncurses=6.1=h0a44026_1
nltk=3.2.4=py36_0
oauthlib=3.0.1=py_0
openssl=1.1.1=h1de35cc_0
pip=19.0.3=py36_0
pycparser=2.19=py36_1
pyjwt=1.7.1=py_0
pymongo=3.7.2=py36h0a44026_0
pyopenssl=19.0.0=py36_0
pysocks=1.6.8=py36_1002
python=3.6.8=haf84260_0
readline=7.0=h1de35cc_5
requests=2.21.0=py36_1000
requests-oauthlib=1.2.0=py_0
setuptools=41.0.0=py36_0
six=1.12.0=py36_1000
sqlite=3.27.2=ha441bb4_0
textblob=0.15.3=py_0
tk=8.6.8=ha441bb4_0
tweepy=3.6.0=py36_0
urllib3=1.24.1=py36_1000
wheel=0.33.1=py36_0
xz=5.2.4=h1de35cc_4
zlib=1.2.11=h1de35cc_3
